# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a wrapper application on alibaba hotel service",
        "title": "Hotel Engine Api",
        "termsOfService": "NONE",
        "contact": {
            "name": "info@jabama.com",
            "url": "jabama.com",
            "email": "info@jabama.com"
        },
        "license": {
            "name": "NONE",
            "url": "None"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/v1/hotel/sync-cities/{secret}": {
            "put": {
                "description": "sync all cities",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel - management"
                ],
                "summary": "sync cities",
                "operationId": "sync-cities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the sync secret",
                        "name": "secret",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CityDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
...
        "/v1/load-inner-types": {
            "get": {
                "description": "this method does not do anything. it just contains the inner objects of other method's responses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "this method does not do anything. it just contains the inner objects of other method's responses",
                "operationId": "LoadInnerTypes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserName",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AvailableRoomDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HotelPlaceDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.RoomOptionDto"
                        }
                    },
                    "402": {
                        "description": "Payment Required",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchDateDto"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchPriceDto"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchScoreDto"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/dto.FilterOutput"
                        }
                    },
                    "406": {
                        "description": "Not Acceptable",
                        "schema": {
                            "$ref": "#/definitions/dto.FilterDto"
                        }
                    },
                    "407": {
                        "description": "Proxy Authentication Required",
                        "schema": {
                            "$ref": "#/definitions/dto.AvailableRoomAdultDto"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/dto.AvailableRoomChildrenDto"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.AvailableRoomPassportDto"
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchResponseHotelDto"
                        }
                    },
                    "411": {
                        "description": "Length Required",
                        "schema": {
                            "$ref": "#/definitions/dto.RequestRoomDto"
                        }
                    },
                    "412": {
                        "description": "Precondition Failed",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchResponseLocationDto"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchResponseRateReviewDto"
                        }
                    },
                    "414": {
                        "description": "Request URI Too Long",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchResponseLocationGeoDto"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/dto.RoomDto"
                        }
                    },
                    "416": {
                        "description": "Requested Range Not Satisfiable",
                        "schema": {
                            "$ref": "#/definitions/dto.MealPlanTypeDto"
                        }
                    },
                    "417": {
                        "description": "Expectation Failed",
                        "schema": {
                            "$ref": "#/definitions/dto.OptionInfoDetailDto"
                        }
                    },
                    "418": {
                        "description": "I'm a teapot",
                        "schema": {
                            "$ref": "#/definitions/dto.OptionInfoDetailRoomDto"
                        }
                    },
                    "419": {
                        "schema": {
                            "$ref": "#/definitions/dto.OptionInfoDetailRestrictedMarkupDto"
                        }
                    },
                    "420": {
                        "schema": {
                            "$ref": "#/definitions/dto.OptionCancellationDto"
                        }
                    },
                    "421": {
                        "description": "Misdirected Request",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderDetailRoomDto"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.HotelDto"
                        }
                    },
                    "423": {
                        "description": "Locked",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderEnquiryItemDto"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderEnquiryItemOptionDto"
                        }
                    },
                    "425": {
                        "description": "Too Early",
                        "schema": {
                            "$ref": "#/definitions/dto.OrderEnquiryItemOptionPassengerInformation"
                        }
                    },
                    "426": {
                        "description": "Upgrade Required",
                        "schema": {
                            "$ref": "#/definitions/dto.HotelSyncDetail"
                        }
                    },
                    "427": {
                        "schema": {
                            "$ref": "#/definitions/dto.HotelBadgeDto"
                        }
                    },
                    "428": {
                        "description": "Precondition Required",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchResponseBadgeDto"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/dto.HotelSeoDto"
                        }
                    },
                    "430": {
                        "schema": {
                            "$ref": "#/definitions/dto.HotelFAQDto"
                        }
                    },
                    "431": {
                        "description": "Request Header Fields Too Large",
                        "schema": {
                            "$ref": "#/definitions/dto.HotelFAQDetailsDto"
                        }
                    }
                }
            }
        },
...
        "/v1/public/amenity-category": {
            "get": {
                "description": "delete amenity categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public - Amenity Category"
                ],
                "summary": "get amenity categories",
                "operationId": "GetAmenityCategories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AmenityCategoriesResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "update amenity category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public - Amenity Category"
                ],
                "summary": "update amenity category",
                "operationId": "UpdateAmenityCategory",
                "parameters": [
                    {
                        "description": "update category dto",
                        "name": "catDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AmenityCategoryDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AmenityCategoryDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            },
            "post": {
                "description": "create amenity category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public - Amenity Category"
                ],
                "summary": "create amenity category",
                "operationId": "CreateAmenityCategory",
                "parameters": [
                    {
                        "description": "create category dto",
                        "name": "catDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AmenityCategoryDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AmenityCategoryDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/public/amenity-category/{id}": {
            "get": {
                "description": "get amenity category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public - Amenity Category"
                ],
                "summary": "get amenity category",
                "operationId": "GetAmenityCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AmenityCategoryDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete amenity category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public - Amenity Category"
                ],
                "summary": "delete amenity category",
                "operationId": "DeleteAmenityCategory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteAmenityCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
...
        "/v1/public/set-badge-icon": {
            "put": {
                "description": "Update badge icon",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public - Badges"
                ],
                "summary": "Update badge icon",
                "operationId": "SetBadgeIcon",
                "parameters": [
                    {
                        "description": "the request body",
                        "name": "badgeIconDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SetBadgeIconDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HotelBadgeDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AvailableRoomDto": {
            "type": "object",
            "properties": {
                "adults": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AvailableRoomAdultDto"
                    }
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AvailableRoomChildrenDto"
                    }
                }
            }
        },
...
        "dto.CityDto": {
            "type": "object",
            "properties": {
                "baseId": {
                    "type": "integer"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "dto.HotelPlaceDto": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "geo_location": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SearchDateDto": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
        "dto.SearchPriceDto": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SearchScoreDto": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
...
        "indraframework.IndraException": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "severity": {
                    "type": "integer"
                },
                "technicalMessage": {
                    "type": "string"
                }
            }
        }
    },
...
}